editor/libeditor/tests/test_nsITableEditor_insertTableCell.html

<!DOCTYPE>
<html>
<head>
  <title>Test for nsITableEditor.insertTableCell()</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<div id="display">
</div>
<div id="content" contenteditable>out of table<table><tr><td>default content</td></tr></table></div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(async () => {
  await SpecialPowers.pushPrefEnv({
    set: [["dom.input_events.beforeinput.enabled", true]],
  });

  let editor = document.getElementById("content");
  let selection = document.getSelection();

  function checkInputEvent(aEvent, aDescription) {
    ok(aEvent instanceof InputEvent,
       `"${aEvent.type}" event should be dispatched with InputEvent interface ${aDescription}`);
    is(aEvent.cancelable, false,
       `"${aEvent.type}" event should be never cancelable ${aDescription}`);
    is(aEvent.bubbles, true,
       `"${aEvent.type}" event should always bubble ${aDescription}`);
    is(aEvent.inputType, "",
       `inputType of "${aEvent.type}" event should be empty string ${aDescription}`);
    is(aEvent.data, null,
       `data of "${aEvent.type}" event should be null ${aDescription}`);
    is(aEvent.dataTransfer, null,
       `dataTransfer of "${aEvent.type}" event should be null ${aDescription}`);
  }

  let beforeInputEvents = [];
  let inputEvents = [];
  function onBeforeInput(aEvent) {
    beforeInputEvents.push(aEvent);
  }
  function onInput(aEvent) {
    inputEvents.push(aEvent);
  }
  editor.addEventListener("beforeinput", onBeforeInput);
  editor.addEventListener("input", onInput);

  beforeInputEvents = [];
  inputEvents = [];
  selection.collapse(editor.firstChild, 0);
  getTableEditor().insertTableCell(1, false);
  is(editor.innerHTML, "out of table<table><tbody><tr><td>default content</td></tr></tbody></table>",
     "nsITableEditor.insertTableCell(1, false) should do nothing if selection is not in <table>");
  is(beforeInputEvents.length, 1,
     '"beforeinput" event should be fired when a call of nsITableEditor.insertTableCell(1, false) even though it will do nothing');
  checkInputEvent(beforeInputEvents[0], "when selection collapsed outside table element (nsITableEditor.insertTableCell(1, false))");
  is(inputEvents.length, 0,
     'No "input" event should be fired when a call of nsITableEditor.insertTableCell(1, false) does nothing');

  beforeInputEvents = [];
  inputEvents = [];
  getTableEditor().insertTableCell(1, true);
  is(editor.innerHTML, "out of table<table><tbody><tr><td>default content</td></tr></tbody></table>",
     "nsITableEditor.insertTableCell(1, true) should do nothing if selection is not in <table>");
  todo_is(beforeInputEvents.length, 1,
     '"beforeinput" event should be fired when a call of nsITableEditor.insertTableCell(1, true) even though it will  do nothing');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection is collapsed outside table element (nsITableEditor.insertTableCell(1, true))");
  }
  is(inputEvents.length, 0,
     'No "input" event should be fired when a call of nsITableEditor.insertTableCell(1, true) does nothing');

  selection.removeAllRanges();
  try {
    beforeInputEvents = [];
    inputEvents = [];
    getTableEditor().insertTableCell(1, false);
    ok(false, "getTableEditor().insertTableCell(1, false) without selection ranges should throw exception");
  } catch (e) {
    ok(true, "getTableEditor().insertTableCell(1, false) without selection ranges should throw exception");
    is(beforeInputEvents.length, 0,
       'No "beforeinput" event should be fired when nsITableEditor.insertTableCell(1, false) causes exception due to no selection range');
    is(inputEvents.length, 0,
       'No "input" event should be fired when nsITableEditor.insertTableCell(1, false) causes exception due to no selection range');
  }
  try {
    beforeInputEvents = [];
    inputEvents = [];
    getTableEditor().insertTableCell(1, true);
    ok(false, "getTableEditor().insertTableCell(1, true) without selection ranges should throw exception");
  } catch (e) {
    ok(true, "getTableEditor().insertTableCell(1, true) without selection ranges should throw exception");
    is(beforeInputEvents.length, 0,
       'No "beforeinput" event should be fired when nsITableEditor.insertTableCell(1, true) causes exception due to no selection range');
    is(inputEvents.length, 0,
       'No "input" event should be fired when nsITableEditor.insertTableCell(1, true) causes exception due to no selection range');
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       "<tr><td>cell1-1</td><td>cell1-2</td></tr>" +
                       '<tr><td id="select">cell2-1</td><td>cell2-2</td></tr>' +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 0);
  getTableEditor().insertTableCell(1, false);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-1</td><td>cell1-2</td></tr>" +
                         '<tr><td valign="top"><br></td><td id="select">cell2-1</td><td>cell2-2</td></tr>' +
                         "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(1, false) should insert a cell before the cell containing selection");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell in middle row (before)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell in middle row (before)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell in middle row (before)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell in middle row (before)");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       "<tr><td>cell1-1</td><td>cell1-2</td></tr>" +
                       '<tr><td id="select">cell2-1</td><td>cell2-2</td></tr>' +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 0);
  getTableEditor().insertTableCell(1, true);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-1</td><td>cell1-2</td></tr>" +
                         '<tr><td id="select">cell2-1</td><td valign="top"><br></td><td>cell2-2</td></tr>' +
                         "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(1, true) should insert a cell after the cell containing selection");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell in middle row (after)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell in middle row (after)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell in middle row (after)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell in middle row (after)");
  }

  // with rowspan.

  // Odd case. This puts the cell containing selection moves right of row-spanning cell.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td rowspan="2">cell1-2</td></tr>' +
                       '<tr><td id="select">cell2-1</td></tr>' +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 0);
  getTableEditor().insertTableCell(1, false);
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td>cell1-1</td><td rowspan="2">cell1-2</td></tr>' +
                         '<tr><td valign="top"><br></td><td id="select">cell2-1</td></tr>' +
                         "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(1, false) should insert a cell before the cell containing selection and moves the cell to right of the row-spanning cell element");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell in middle row and it has row-spanned cell (before)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell in middle row and it has row-spanned cell (before)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell in middle row and it has row-spanned cell (before)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell in middle row and it has row-spanned cell (before)");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td rowspan="3">cell1-2</td></tr>' +
                       '<tr><td id="select">cell2-1</td></tr>' +
                       "<tr><td>cell3-1</td></tr>" +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 0);
  getTableEditor().insertTableCell(1, true);
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td>cell1-1</td><td rowspan="3">cell1-2</td></tr>' +
                         '<tr><td id="select">cell2-1</td><td valign="top"><br></td></tr>' +
                         "<tr><td>cell3-1</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(1, true) should insert a cell after the cell containing selection and moves the cell to right of the row-spanning cell element");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell in middle row and it has row-spanned cell (after)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell in middle row and it has row-spanned cell (after)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell in middle row and it has row-spanned cell (after)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell in middle row and it has row-spanned cell (after)");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td id="select" rowspan="2">cell1-2</td></tr>' +
                       "<tr><td>cell2-1</td></tr>" +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 1);
  getTableEditor().insertTableCell(2, false);
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td>cell1-1</td><td valign="top"><br></td><td valign="top"><br></td><td id="select" rowspan="2">cell1-2</td></tr>' +
                         "<tr><td>cell2-1</td></tr>" +
                         "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(2, false) should insert 2 cells before the row-spanning cell containing selection");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell in row-spanning (before)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell in row-spanning (before)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell in row-spanning (before)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell in row-spanning (before)");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td id="select" rowspan="2">cell1-2</td></tr>' +
                       "<tr><td>cell2-1</td></tr>" +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 1);
  getTableEditor().insertTableCell(2, true);
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td>cell1-1</td><td id="select" rowspan="2">cell1-2</td><td valign="top"><br></td><td valign="top"><br></td></tr>' +
                         "<tr><td>cell2-1</td></tr>" +
                         "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(2, false) should insert 2 cells after the row-spanning cell containing selection");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell in row-spanning (after)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell in row-spanning (after)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell in row-spanning (after)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell in row-spanning (after)");
  }

  // with colspan

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td id="select">cell1-2</td><td>cell1-3</td></tr>' +
                       '<tr><td colspan="2">cell2-1</td><td>cell2-3</td></tr>' +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 0);
  getTableEditor().insertTableCell(1, false);
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td>cell1-1</td><td valign="top"><br></td><td id="select">cell1-2</td><td>cell1-3</td></tr>' +
                         '<tr><td colspan="2">cell2-1</td><td>cell2-3</td></tr>' +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(1, false) should insert a cell before the cell containing selection but do not modify col-spanning cell");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell whose next row cell is col-spanned (before)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell whose next row cell is col-spanned (before)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell whose next row cell is col-spanned (before)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell whose next row cell is col-spanned (before)");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td id="select">cell1-2</td><td>cell1-3</td></tr>' +
                       '<tr><td colspan="3">cell2-1</td></tr>' +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 0);
  getTableEditor().insertTableCell(1, true);
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td>cell1-1</td><td id="select">cell1-2</td><td valign="top"><br></td><td>cell1-3</td></tr>' +
                         '<tr><td colspan="3">cell2-1</td></tr>' +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(1, true) should insert a cell after the cell containing selection but do not modify col-spanning cell");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell whose next row cell is col-spanned (after)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell whose next row cell is col-spanned (after)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell whose next row cell is col-spanned (after)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell whose next row cell is col-spanned (after)");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       "<tr><td>cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>" +
                       '<tr><td id="select" colspan="2">cell2-1</td><td>cell2-3</td></tr>' +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 1);
  getTableEditor().insertTableCell(2, false);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>" +
                         '<tr><td valign="top"><br></td><td valign="top"><br></td><td id="select" colspan="2">cell2-1</td><td>cell2-3</td></tr>' +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(2, false) should insert 2 cells before the col-spanning cell containing selection");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell which is col-spanning (before)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell which is col-spanning (before)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell which is col-spanning (before)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell which is col-spanning (before)");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       "<tr><td>cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>" +
                       '<tr><td id="select" colspan="2">cell2-1</td><td>cell2-3</td></tr>' +
                     "</table>";
  editor.focus();
  editor.scrollTop; // layout information required.
  beforeInputEvents = [];
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 1);
  getTableEditor().insertTableCell(2, true);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>" +
                         '<tr><td id="select" colspan="2">cell2-1</td><td valign="top"><br></td><td valign="top"><br></td><td>cell2-3</td></tr>' +
                       "</tbody></table>",
     "nsITableEditor.insertTableCell(2, false) should insert 2 cells after the col-spanning cell containing selection");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when selection collapsed in a cell which is col-spanning (after)');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection collapsed in a cell which is col-spanning (after)");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when selection collapsed in a cell which is col-spanning (after)');
  if (inputEvents.length > 0) {
    checkInputEvent(inputEvents[0], "when selection collapsed in a cell which is col-spanning (after)");
  }

  editor.removeEventListener("beforeinput", onBeforeInput);
  editor.removeEventListener("input", onInput);

  SimpleTest.finish();
});

function getTableEditor() {
  var editingSession = SpecialPowers.wrap(window).docShell.editingSession;
  return editingSession.getEditorForWindow(window).QueryInterface(SpecialPowers.Ci.nsITableEditor);
}

</script>
</body>

</html>
